+++ src/control_plane/workload/mod.rs
@@ -10,6 +10,7 @@ mod immutable_project_application_plan;
 mod javascript_runtime_spec;
 mod node_package_manager;
 mod orphaned_project_workload_options;
+mod php_tool;
 mod plan_immutable_project_application;
 mod plan_project_process_operation;
 mod project_command;
@@ -49,6 +50,7 @@ pub(crate) use immutable_project_application_plan::ImmutableProjectApplicationPl
 pub(crate) use javascript_runtime_spec::JavaScriptRuntimeSpec;
 pub(crate) use node_package_manager::NodePackageManager;
 pub(crate) use orphaned_project_workload_options::OrphanedProjectWorkloadOptions;
+pub(crate) use php_tool::PhpTool;
 pub(crate) use plan_immutable_project_application::plan_immutable_project_application;
 pub(crate) use plan_project_process_operation::plan_project_process_operation;
 pub(crate) use project_command::ProjectCommand;